fix: 카카오메시지 자료형 수정 (#35)

+++ src/main/java/com/knu/daeguhackathon/kakaotoken/dto/KakaoResultRequest.java
@@ -1,6 +1,6 @@
 package com.knu.daeguhackathon.kakaotoken.dto;
 public record KakaoResultRequest(
-    String start, String end, int estimatedTime, Double buildingDistance
+    String start, String end, Double estimatedTime, Double buildingDistance
 ) {
 
 }
